refactor: Move overlays to 'Overlay'
