+++ web/pages/register.tsx
@@ -16,7 +16,7 @@ import { FaKey } from 'react-icons/fa';
 
 import TextInput from '../components/Input/TextInput';
 import Layout from '../components/Layout';
-import QRDialog from '../components/QRDialog';
+import QRDialog from '../components/Overlay/QRDialog';
 import axios from '../utils/http';
 import routes from '../utils/routes';
 import type { User } from '../utils/types';
